move and merge
