+++ web/apps/photos/src/components/Upload/UploadProgress/dialog.tsx
@@ -1,7 +1,10 @@
 import { Dialog, DialogContent, Link } from "@mui/material";
 import { t } from "i18next";
 
-import { UPLOAD_RESULT, UPLOAD_STAGES } from "@/new/photos/constants/upload";
+import {
+    UPLOAD_RESULT,
+    UPLOAD_STAGES,
+} from "@/new/photos/services/upload/types";
 import { dialogCloseHandler } from "@ente/shared/components/DialogBox/TitleWithCloseButton";
 import UploadProgressContext from "contexts/uploadProgress";
 import { useContext, useEffect, useState } from "react";
