+++ web/apps/photos/src/components/Upload/UploadProgress/progressBar.tsx
@@ -1,4 +1,4 @@
-import { UPLOAD_STAGES } from "@/new/photos/constants/upload";
+import { UPLOAD_STAGES } from "@/new/photos/services/upload/types";
 import { Box, Divider, LinearProgress } from "@mui/material";
 import UploadProgressContext from "contexts/uploadProgress";
 import { useContext } from "react";
