+++ web/apps/photos/src/components/Upload/UploadProgress/title.tsx
@@ -1,4 +1,4 @@
-import { UPLOAD_STAGES } from "@/new/photos/constants/upload";
+import { UPLOAD_STAGES } from "@/new/photos/services/upload/types";
 import {
     IconButtonWithBG,
     SpaceBetweenFlex,
