+++ web/apps/photos/src/services/upload/uploadManager.ts
@@ -1,11 +1,12 @@
 import { FILE_TYPE } from "@/media/file-type";
 import { potentialFileTypeFromExtension } from "@/media/live-photo";
+import { getLocalFiles } from "@/new/photos/services/files";
+import type { UploadItem } from "@/new/photos/services/upload/types";
 import {
     RANDOM_PERCENTAGE_PROGRESS_FOR_PUT,
     UPLOAD_RESULT,
     UPLOAD_STAGES,
-} from "@/new/photos/constants/upload";
-import { getLocalFiles } from "@/new/photos/services/files";
+} from "@/new/photos/services/upload/types";
 import { EncryptedEnteFile, EnteFile } from "@/new/photos/types/file";
 import { ensureElectron } from "@/next/electron";
 import { lowercaseExtension, nameAndExtension } from "@/next/file";
@@ -35,7 +36,6 @@ import {
     tryParseTakeoutMetadataJSON,
     type ParsedMetadataJSON,
 } from "./takeout";
-import type { UploadItem } from "./types";
 import UploadService, { uploadItemFileName, uploader } from "./uploadService";
 
 export type FileID = number;
