+++ web/packages/new/photos/services/upload/types.ts
@@ -1,4 +1,5 @@
 import type { ZipItem } from "@/next/types/ipc";
+import type { Location } from "../../types/metadata";
 
 /**
  * An item to upload is one of the following:
@@ -55,3 +56,28 @@ export const toDataOrPathOrZipEntry = (desktopUploadItem: DesktopUploadItem) =>
     typeof desktopUploadItem == "string" || Array.isArray(desktopUploadItem)
         ? desktopUploadItem
         : desktopUploadItem.path;
+
+export const RANDOM_PERCENTAGE_PROGRESS_FOR_PUT = () => 90 + 10 * Math.random();
+
+export const NULL_LOCATION: Location = { latitude: null, longitude: null };
+
+export enum UPLOAD_STAGES {
+    START,
+    READING_GOOGLE_METADATA_FILES,
+    EXTRACTING_METADATA,
+    UPLOADING,
+    CANCELLING,
+    FINISH,
+}
+
+export enum UPLOAD_RESULT {
+    FAILED,
+    ALREADY_UPLOADED,
+    UNSUPPORTED,
+    BLOCKED,
+    TOO_LARGE,
+    LARGER_THAN_AVAILABLE_STORAGE,
+    UPLOADED,
+    UPLOADED_WITH_STATIC_THUMBNAIL,
+    ADDED_SYMLINK,
+}
